+++ InnerEye/ML/configs/histo_configs/classification/DeepSMILEPanda.py
@@ -3,28 +3,22 @@
 #  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 #  ------------------------------------------------------------------------------------------
 
-"""DeepSMILECrck is the container for experiments relating to DeepSMILE using the TCGA-CRCk dataset.
-Run using `python InnerEyePrivate/ML/runner.py --model=DeepSMILECrck --encoder_type=<encoder class name>`
-
-For convenience, this module also defines encoder-specific containers that can be invoked without
-additional arguments, e.g. `python InnerEyePrivate/ML/runner.py --model=TcgaCrckImageNetMIL`
-
-Reference:
-- Schirris (2021). DeepSMILE: Self-supervised heterogeneity-aware multiple instance learning for DNA
-damage response defect classification directly from H&E whole-slide images. arXiv:2107.09405
-"""
-from pathlib import Path
 from typing import Any, Dict
-
+from pathlib import Path
+import os
 from monai.transforms import Compose
 from pytorch_lightning.callbacks.model_checkpoint import ModelCheckpoint
 
+from health_azure.utils import CheckpointDownloader
+from health_azure.utils import get_workspace
 from health_ml.networks.layers.attention_layers import GatedAttentionLayer
 from InnerEye.Common import fixed_paths
-from InnerEye.ML.configs.histo_configs.classification.BaseMIL import BaseMIL
-from InnerEye.ML.Histopathology.datamodules.base_module import TilesDataModule
-from InnerEye.ML.Histopathology.datamodules.tcga_crck_module import (
-    TcgaCrckTilesDataModule,
+from InnerEye.ML.Histopathology.datamodules.panda_module import PandaTilesDataModule
+from InnerEye.ML.Histopathology.datasets.panda_tiles_dataset import PandaTilesDataset
+
+from InnerEye.ML.Histopathology.models.transforms import (
+    EncodeTilesBatchd,
+    LoadTilesBatchd,
 )
 from InnerEye.ML.Histopathology.models.encoders import (
     HistoSSLEncoder,
@@ -32,28 +26,22 @@ from InnerEye.ML.Histopathology.models.encoders import (
     ImageNetSimCLREncoder,
     InnerEyeSSLEncoder,
 )
-from InnerEye.ML.Histopathology.models.transforms import (
-    EncodeTilesBatchd,
-    LoadTilesBatchd,
-)
-from InnerEye.ML.Histopathology.datasets.tcga_crck_tiles_dataset import (
-    TcgaCrck_TilesDataset,
-)
+from InnerEye.ML.configs.histo_configs.classification.BaseMIL import BaseMIL
+from InnerEye.ML.configs.histo_configs.run_ids import innereye_ssl_checkpoint
 
 
-class DeepSMILECrck(BaseMIL):
+class DeepSMILEPanda(BaseMIL):
     def __init__(self, **kwargs: Any) -> None:
-        # Define dictionary with default params that can be overriden from subclasses or CLI
         default_kwargs = dict(
             # declared in BaseMIL:
             pooling_type=GatedAttentionLayer.__name__,
             # declared in DatasetParams:
-            local_dataset=Path("/tmp/datasets/TCGA-CRCk"),
-            azure_dataset_id="TCGA-CRCk",
+            local_dataset=Path("/tmp/datasets/PANDA_tiles"),
+            azure_dataset_id="PANDA_tiles",
             # To mount the dataset instead of downloading in AML, pass --use_dataset_mount in the CLI
             # declared in TrainerParams:
-            num_epochs=16,
-            recovery_checkpoint_save_interval=16,
+            num_epochs=200,
+            recovery_checkpoint_save_interval=10,
             recovery_checkpoints_save_last_k=-1,
             # declared in WorkflowParams:
             number_of_cross_validation_splits=5,
@@ -65,16 +53,15 @@ class DeepSMILECrck(BaseMIL):
         )
         default_kwargs.update(kwargs)
         super().__init__(**default_kwargs)
-
+        super().__init__(**default_kwargs)
         self.best_checkpoint_filename = "checkpoint_max_val_auroc"
         self.best_checkpoint_filename_with_suffix = (
             self.best_checkpoint_filename + ".ckpt"
         )
         self.checkpoint_folder_path = "outputs/checkpoints/"
-
         best_checkpoint_callback = ModelCheckpoint(
             dirpath=self.checkpoint_folder_path,
-            monitor="val/auroc",
+            monitor="val/accuracy",
             filename=self.best_checkpoint_filename,
             auto_insert_metric_name=False,
             mode="max",
@@ -87,15 +74,29 @@ class DeepSMILECrck(BaseMIL):
             f"/tmp/innereye_cache1/{self.__class__.__name__}-{self.encoder_type}/"
         )
 
-    def get_data_module(self) -> TilesDataModule:
-        image_key = TcgaCrck_TilesDataset.IMAGE_COLUMN
+    def setup(self) -> None:
+        if self.encoder_type == InnerEyeSSLEncoder.__name__:
+            self.downloader = CheckpointDownloader(
+                azure_config_json_path=get_workspace(),
+                run_recovery_id=innereye_ssl_checkpoint,
+                checkpoint_filename="last.ckpt",
+                download_dir="outputs/",
+            )
+            os.chdir(fixed_paths.repository_root_directory())
+            self.downloader.download_checkpoint_if_necessary()
+        self.encoder = self.get_encoder()
+        self.encoder.cuda()
+        self.encoder.eval()
+
+    def get_data_module(self) -> PandaTilesDataModule:
+        image_key = PandaTilesDataset.IMAGE_COLUMN
         transform = Compose(
             [
                 LoadTilesBatchd(image_key, progress=True),
                 EncodeTilesBatchd(image_key, self.encoder),
             ]
         )
-        return TcgaCrckTilesDataModule(
+        return PandaTilesDataModule(
             root_path=self.local_dataset,
             max_bag_size=self.max_bag_size,
             batch_size=self.batch_size,
@@ -124,21 +125,21 @@ class DeepSMILECrck(BaseMIL):
         )
 
 
-class TcgaCrckImageNetMIL(DeepSMILECrck):
+class PandaImageNetMIL(DeepSMILEPanda):
     def __init__(self, **kwargs: Any) -> None:
         super().__init__(encoder_type=ImageNetEncoder.__name__, **kwargs)
 
 
-class TcgaCrckImageNetSimCLRMIL(DeepSMILECrck):
+class PandaImageNetSimCLRMIL(DeepSMILEPanda):
     def __init__(self, **kwargs: Any) -> None:
         super().__init__(encoder_type=ImageNetSimCLREncoder.__name__, **kwargs)
 
 
-class TcgaCrckInnerEyeSSLMIL(DeepSMILECrck):
+class PandaInnerEyeSSLMIL(DeepSMILEPanda):
     def __init__(self, **kwargs: Any) -> None:
         super().__init__(encoder_type=InnerEyeSSLEncoder.__name__, **kwargs)
 
 
-class TcgaCrckHistoSSLMIL(DeepSMILECrck):
+class PandaHistoSSLMIL(DeepSMILEPanda):
     def __init__(self, **kwargs: Any) -> None:
         super().__init__(encoder_type=HistoSSLEncoder.__name__, **kwargs)