+++ Tests/ML/models/architectures/test_image_encoder_with_mlp.py
@@ -273,6 +273,8 @@ def test_image_encoder_with_segmentation(test_output_dirs: OutputFolderForTests,
                           aggregation_type=aggregation_type,
                           scan_size=scan_size)
     config.use_mixed_precision = True
+    # Necessary because torch otherwise says "avg_pool3d_backward_cuda does not have a deterministic implementation"
+    config.pl_deterministic = False
     config.set_output_to(test_output_dirs.root_dir)
     config.num_epochs = 1
     config.local_dataset = Path()